Use the symbol table to drive the disassembler
This is definitely an idealised case and doesn't take into account
edge cases such as lack of a symbol table or missing symbol sizes
of the symbols in the symbol table. A real nasty one is partially
filled in symbol table where most of the symbols are specified, but
some still very much missing.